refactor use core/enable_if.hpp over utility/enable_if.hpp
